+++ backend/canisters/local_group_index/api/src/lifecycle/init.rs
@@ -1,6 +1,6 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
-use types::{CanisterId, CanisterWasm, Version};
+use types::{CanisterId, CanisterWasm, UserId, Version};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
@@ -16,5 +16,6 @@ pub struct Args {
     pub notifications_canister_id: CanisterId,
     pub cycles_dispenser_canister_id: CanisterId,
     pub ledger_canister_id: CanisterId,
+    pub proposals_bot_user_id: UserId,
     pub test_mode: bool,
 }